add eco lett params, cool Hopf (?) in Ft (disequi)
I've been playing with the length of the time series relative to the range of qE to explore, and if time series is short relative to range(qE), the system is more out of equilibrium. Leads to some cooler looking dynamics.

I wonder if S-map etc would work if the system was not at equilibrium.

Ft (planktivores) show what looks like a Hopf bifurcation in the current saved form.