create a "function" for FW parameters
created to make running the model more efficient – don't have to redefine parameters that you wouldn't change

however, this isn't really a "function". I could do something to assign in global environment, perhaps, but right now it's just a script. I want to think of a better way to do this.